Add block template generation function
